Tweaks to how login/search works.
So in order for us to handle submit forms, we're using our Servlets. We
do not need doGet or doPost as we can just use the callbacks to handle
onClick or onSelection methods. I've introduced a test scenario for
logining in which simply creates a new user on login temporary.